+++ spec/identity_registration_spec.rb
@@ -23,43 +23,43 @@ describe 'openstack-block-storage::identity_registration' do
       )
     end
 
-    it 'registers cinder volume service' do
+    it 'registers cinder v2 volume service' do
       expect(chef_run).to create_service_openstack_identity_register(
-        'Register Cinder Volume Service'
+        'Register Cinder V2 Volume Service'
       ).with(
         auth_uri: 'http://127.0.0.1:35357/v2.0',
         bootstrap_token: 'bootstrap-token',
-        service_name: 'cinder',
-        service_type: 'volume',
-        service_description: 'Cinder Volume Service',
+        service_name: 'cinderv2',
+        service_type: 'volumev2',
+        service_description: 'Cinder Volume Service V2',
         endpoint_region: 'RegionOne',
-        endpoint_adminurl: 'http://127.0.0.1:8776/v1/%(tenant_id)s',
-        endpoint_internalurl: 'http://127.0.0.1:8776/v1/%(tenant_id)s',
-        endpoint_publicurl: 'http://127.0.0.1:8776/v1/%(tenant_id)s'
+        endpoint_adminurl: 'http://127.0.0.1:8776/v2/%(tenant_id)s',
+        endpoint_internalurl: 'http://127.0.0.1:8776/v2/%(tenant_id)s',
+        endpoint_publicurl: 'http://127.0.0.1:8776/v2/%(tenant_id)s'
       )
     end
 
-    context 'registers volume endpoint' do
+    context 'registers v2 volume endpoint' do
       it 'with default values' do
         expect(chef_run).to create_endpoint_openstack_identity_register(
-          'Register Cinder Volume Endpoint'
+          'Register Cinder V2 Volume Endpoint'
         ).with(
           auth_uri: 'http://127.0.0.1:35357/v2.0',
           bootstrap_token: 'bootstrap-token',
-          service_name: 'cinder',
-          service_type: 'volume',
-          service_description: 'Cinder Volume Service',
+          service_name: 'cinderv2',
+          service_type: 'volumev2',
+          service_description: 'Cinder Volume Service V2',
           endpoint_region: 'RegionOne',
-          endpoint_adminurl: 'http://127.0.0.1:8776/v1/%(tenant_id)s',
-          endpoint_internalurl: 'http://127.0.0.1:8776/v1/%(tenant_id)s',
-          endpoint_publicurl: 'http://127.0.0.1:8776/v1/%(tenant_id)s'
+          endpoint_adminurl: 'http://127.0.0.1:8776/v2/%(tenant_id)s',
+          endpoint_internalurl: 'http://127.0.0.1:8776/v2/%(tenant_id)s',
+          endpoint_publicurl: 'http://127.0.0.1:8776/v2/%(tenant_id)s'
         )
       end
 
       it 'with custom region override' do
         node.set['openstack']['block-storage']['region'] = 'volumeRegion'
         expect(chef_run).to create_endpoint_openstack_identity_register(
-          'Register Cinder Volume Endpoint'
+          'Register Cinder V2 Volume Endpoint'
         ).with(endpoint_region: 'volumeRegion')
       end
     end